requirements.txt
langchain==0.0.249
matplotlib==3.5.2
openai==0.27.8
streamlit==1.25.0
textstat==0.7.3
pyperclip==1.8.2
azure-storage-blob==12.18.0b1
streamlit-authenticator==0.2.2
statistics